v2.23.7

+++ src/client/client-common/ipc/server.cpp
@@ -1,6 +1,7 @@
 #include "server.h"
 #include "connection.h"
 #include "utils/log/categories.h"
+#include <QDir>
 #include <QFile>
 #include <QStandardPaths>
 
@@ -18,18 +19,25 @@ Server::~Server()
 
 bool Server::start()
 {
+    server_.setSocketOptions(QLocalServer::UserAccessOption);
 
 #if defined(Q_OS_MACOS)
-    const QString runtimeDir = QStandardPaths::writableLocation(QStandardPaths::TempLocation);
-    const QString path = runtimeDir + "/windscribe-localipc.sock";
+    const QString runtimeDir = QStandardPaths::writableLocation(QStandardPaths::GenericDataLocation) +
+        "/" WS_SETTINGS_ORG "/ipc";
+    QDir runtimeDirObj(runtimeDir);
+    if (!runtimeDirObj.mkpath(".")) {
+        qCCritical(LOG_IPC) << "IPC server could not create runtime directory:" << runtimeDir;
+        return false;
+    }
+    QFile::setPermissions(runtimeDir,
+        QFileDevice::ReadOwner | QFileDevice::WriteOwner | QFileDevice::ExeOwner);
+    const QString path = runtimeDir + "/" WS_PRODUCT_NAME_LOWER "-localipc.sock";
     QFile::remove(path);
-    server_.setSocketOptions(QLocalServer::UserAccessOption);
     bool b = server_.listen(path);
 #elif defined(Q_OS_LINUX)
     const QString runtimeDir = QStandardPaths::writableLocation(QStandardPaths::RuntimeLocation);
-    const QString path = runtimeDir + "/windscribe-localipc.sock";
+    const QString path = runtimeDir + "/" WS_PRODUCT_NAME_LOWER "-localipc.sock";
     QFile::remove(path);
-    server_.setSocketOptions(QLocalServer::UserAccessOption);
     bool b = server_.listen(path);
 #else
     bool b = server_.listen(WS_APP_IDENTIFIER "8rM7bza5OR");
